Combine test functions
